Modularize tirestino (draw functions, constants, types)
