+++ src/components/starter/header/header.module.css
@@ -19,15 +19,20 @@
   font-size: 0.7rem;
 }
 
-.header li a {
+.header li a,
+.header h3 a {
   color: white;
   display: inline-block;
   padding: 0;
   text-decoration: none;
+
+  transition: color 300ms ease-in-out;
 }
 
 .header li a:hover {
   color: var(--qwik-light-blue);
+
+  transition: color 200ms ease-in-out;
 }
 
 @media (min-width: 450px) {
