+++ src/tirest/constants.ts
@@ -1,0 +1,9 @@
+import type { Size } from './types'
+
+export const INPUT_DELAY_MS: number = 45
+export const AUTO_FALL_MS: number = 1000
+
+export const fieldSize: Size = {
+  width: 10,
+  height: 20,
+}
